basemaps: move remove to its own view
