+++ lib/assets/javascripts/cartodb3/editor/layers/basemap-content-views/basemap-mosaic-item-view.js
@@ -2,8 +2,7 @@
 
 var template = require('./basemap-mosaic-item.tpl');
 var MosaicItemView = require('../../../components/mosaic/mosaic-item-view');
-var CustomListCollection = require('../../../components/custom-list/custom-list-collection');
-var ContextMenuView = require('../../../components/context-menu/context-menu-view');
+var BasemapMosaicRemoveView = require('./basemap-mosaic-remove-view');
 
 var DEFAULT_SUBDOMAIN = 'a';
 var DEFAULT_X_POSITION = 30;
@@ -14,14 +13,10 @@ var DEFAULT_NAME = _t('editor.layers.basemap.custom-basemap');
 
 module.exports = MosaicItemView.extend({
 
-  events: {
-    'click .js-Mosaic-remove': '_onToggleContextMenuClicked',
-    'mouseenter': '_onMouseEnter',
-    'mouseleave': '_onMouseLeave',
-    'click': '_onClick'
-  },
-
   initialize: function (opts) {
+    if (!opts.basemapsCollection) throw new Error('basemapsCollection is required');
+
+    this._basemapsCollection = opts.basemapsCollection;
     this._disabled = opts.disabled;
 
     this._initBinds();
@@ -36,7 +31,19 @@ module.exports = MosaicItemView.extend({
     );
     this.$el.addClass('js-' + this.model.getValue());
     this.$el.toggleClass('is-selected', !!this.model.get('selected'));
+
+    this._initViews();
+
     return this;
+  },
+
+  _initViews: function () {
+    var basemapMosaicRemoveView = new BasemapMosaicRemoveView({
+      model: this.model,
+      basemapsCollection: this._basemapsCollection
+    });
+    this.addView(basemapMosaicRemoveView);
+    this.$el.append(basemapMosaicRemoveView.render().el);
   },
 
   _getImageURL: function () {
@@ -76,62 +83,6 @@ module.exports = MosaicItemView.extend({
     }
 
     return name;
-  },
-
-  _hasContextMenu: function () {
-    return this._menuView;
-  },
-
-  _hideContextMenu: function () {
-    this._menuView.remove();
-    this.removeView(this._menuView);
-    delete this._menuView;
-  },
-
-  _showContextMenu: function (position) {
-    var menuItems = new CustomListCollection([{
-      label: 'Remove baselayer',
-      val: 'remove-baselayer',
-      destructive: true
-    }]);
-
-    var triggerElementID = 'context-menu-trigger-' + this.model.cid;
-    this.$('.js-Mosaic-remove').attr('id', triggerElementID);
-    this._menuView = new ContextMenuView({
-      collection: menuItems,
-      triggerElementID: triggerElementID,
-      position: position
-    });
-
-    menuItems.bind('change:selected', function (menuItem) {
-      if (menuItem.get('val') === 'remove-baselayer') {
-        this._deleteBaselayer();
-      }
-    }, this);
-
-    this._menuView.model.bind('change:visible', function (model, isContextMenuVisible) {
-      if (this._hasContextMenu() && !isContextMenuVisible) {
-        this._hideContextMenu();
-      }
-    }, this);
-
-    this._menuView.show();
-    this.addView(this._menuView);
-  },
-
-  _deleteBaselayer: function () {
-    console.log('olakase');
-  },
-
-  _onToggleContextMenuClicked: function (e) {
-    if (this._hasContextMenu()) {
-      this._hideContextMenu();
-    } else {
-      this._showContextMenu({
-        x: e.pageX,
-        y: e.pageY
-      });
-    }
   }
 
 });
