+++ lib/assets/javascripts/cartodb3/editor/layers/basemap-content-views/basemap-mosaic-item.tpl
@@ -1,4 +1,3 @@
 <button class="Mosaic-inner">
   <%= template %>
-  <span class="Mosaic-remove js-Mosaic-remove"></span>
 </button>
