+++ lib/assets/javascripts/cartodb3/editor/layers/basemap-content-views/basemap-mosaic-remove.tpl
@@ -1,0 +1,1 @@
+<span class="Mosaic-remove js-Mosaic-remove"></span>
