+++ lib/assets/javascripts/cartodb3/editor/layers/basemap-content-views/basemap-mosaic-view.js
@@ -31,7 +31,8 @@ module.exports = MosaicView.extend({
   _createItem: function (mdl) {
     var view = new BasemapMosaicItemView({
       model: mdl,
-      disabled: this._disabled
+      disabled: this._disabled,
+      basemapsCollection: this._basemapsCollection
     });
     this.$('.js-mosaic').append(view.render().el);
     this.addView(view);
