+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cartodb-ui",
-  "version": "4.2.31",
+  "version": "4.2.32",
   "description": "CartoDB UI frontend",
   "repository": {
     "type": "git",
